+++ src/core/distributions/phylogenetics/Dist_topologyunif.h
@@ -52,7 +52,7 @@ class Dist_topologyunif: public DistributionDiscrete {
         Topology*                   rv(void);                                                           //!< Generate random variable
 
     private:
-        void                        buildRandomBinaryTree(std::vector<TopologyNode*>& tips, int numTaxa);        //!< Build a random binary tree recursively until we have numTaxa tips
+        void                        buildRandomBinaryTree(std::vector<TopologyNode*>& tips, size_t numTaxa);    //!< Build a random binary tree recursively until we have numTaxa tips
         void                        calculateTopologyProb(void);                                        //!< Calculate and set probability member variables
         void                        calculateNumberOfStates(void);                                      //!< Calculate and set number of states member variables
